Added method addToEdgeAttributes
Updates this tree by adding the edge lengths/attributes by the given
coordinates (or their negation). By negation, we mean negative
coordinates become positive, and vice versa. This method does not test
if this would result in any negative edge lengths. Assumes the coords
are in the order that they would be if they were logmap coords with this
tree as the base tree.